==> notebooks/06c_LM-selected-meds.Rmd <==
---
title: "Linear models"
output: html_notebook
---

# Goal
Jacobo de la Cuesta Zuluaga. February 2021.

The goal of this notebook is to fit linear models to dissentangle cardiometabolic disease and obesity associations with the functional profile of the gut metagenome of a sample of adult Colombians.

**Note** that this notebook uses `R v.3.6.2` unlike the other notebooks in the analysis of the Colombian metagenome data, which use `R v.4.0.2`. This is due to issues I encountered while installing `Maaslin2` under `R v.4.0.2`.

# Init
Load packages
```{r}
# Misc
library(LeyLabRMisc)
library(conflicted)

# Graphs
library(cowplot)
library(UpSetR)

# Data manipulation
library(glue)
library(data.table)
library(tidyverse)
library(broom)
#library(furrr)

# Analyis
library(vegan)
library(Maaslin2)
```

```{r}
conflict_prefer("select", "dplyr")
conflict_prefer("filter", "dplyr")
```

Load other functions
```{r}
# JdlC's util functions
source("/ebio/abt3_projects/columbian_gut_metagenome/code/misc_r_functions/misc_jdlc.R")
```

# Var
Load tables and prepare data
```{r}
# Directories
base_dir = "/ebio/abt3_projects/columbian_gut_metagenome"
#LLMGP_dir = file.path(base_dir, "data/metagenome_profile/2019_01_GTDB/humann2")
LLMGP_dir = file.path(base_dir, "data/metagenome_profile/2020_04_combined/humann2")
Feat_map_dir = "/ebio/abt3_projects/databases_no-backup/humann2/utility_mapping"
metadata_dir = file.path(base_dir, "data/metadata")
fig_dir = file.path(base_dir, "code/metagenome_profiles/R_notebooks/Figures")
tab_dir = file.path(base_dir, "code/metagenome_profiles/R_notebooks/Tables")
logratios_dir = file.path(base_dir, "data/metagenome_profile/intermediate_files/2020_04_log_ratios")
annotation_dir = file.path(base_dir, "data/metagenome_profile/intermediate_files/Annotations/")
```

Set seed
```{r}
set.seed(1990)
#plan(multisession, workers = 2)
```


## Load tables
Load metagenome coverage data from nonpareil to determine the samples to retain for analyses. In this case I will use samples with more than 500.000 reads or that have a metagenome coverage above 60%. This according to doi:10.1128/mSystems.00069-18, doi:10.1371/journal.pcbi.1004573 and doi:10.1038/ismej.2014.76
```{r}
nonpareil_table_file = file.path(base_dir, "data/nonpareil/coverage/2020_04_combined_coverage.tsv") 
nonpareil_table = read_tsv(nonpareil_table_file) %>% 
    mutate(Sample = str_replace_all(Sample, "-", "_"))
nonpareil_table %>% dfhead()

Col_retain = nonpareil_table %>% 
    filter(Total.Sequences >= 500000 | C >= 0.6, str_detect(Sample, pattern = "MI")) %>% 
    pull(Sample)
```

Load the metadata of the subjects to be used in differential proportionality and abundance analyses.
```{r}
Col_metadata_raw = file.path(metadata_dir, "Colombia_curated_metadata.tsv") %>% 
    fread(., data.table = F)

Col_metadata = Col_metadata_raw %>% 
    filter(ID %in% Col_retain) %>% 
    mutate(BMI_status = factor(BMI_status, 
                               c("Lean", "Overweight", "Obese")),
           chs_class = factor(chs_class, 
                              c("Healthy", "Abnormal")), 
           bsp_class = factor(bsp_class, 
                              c("Normoweight-Healthy", "Normoweight-Abnormal",
                                "Overweight-Healthy", "Overweight-Abnormal",
                                "Obese-Healthy", "Obese-Abnormal")), 
           Waist_Circumference_status = factor(Waist_Circumference_status, 
                                               c("Normal", "Central_Obesity")), 
           Fat_percentage_status = factor(Fat_percentage_status, 
                                          c("Low", "Normal", "Excess", "Obese")))

# N medications data
Meds_vars = c("Hypertension_med", "Diabetes_med", "Dislipid_med", "PPI_med")
Col_metadata = Col_metadata %>% 
  rowwise() %>% 
  mutate(across(matches("_med"), function(x) ifelse(x == TRUE, 1, 0)), 
         n_med = sum(c_across(Meds_vars))) %>% 
  as.data.frame()
```

Load the clr-transformed functional profile tables. Recall to calculate the clr-transform, I removed the unmapped and the unintegrated features which are technically but not biologically relevant. See CoDa-transformations notebook.
```{r}
# KEGG Modules
## Summary
Col_module_summ = fread(file.path(base_dir,
                                      "code/metagenome_profiles/R_notebooks/Tables/Candidate_Modules.tsv"), 
                                     data.table = FALSE)
## clr-transformed data
Col_module_clr = fread(file.path(logratios_dir, "Candidate_Modules_clr.tsv"), 
                                   data.table = FALSE)

## proportionality matrix
Col_module_rho = fread(file.path(logratios_dir, "Candidate_Modules_rho.tsv"), 
                                   data.table = FALSE)

# KEGG Orthologs
## Summary
Col_ortho_summ = fread(file.path(base_dir,
                                      "code/metagenome_profiles/R_notebooks/Tables/Candidate_Orthologs.tsv"), 
                                    data.table = FALSE)
## clr-transformed data
Col_ortho_clr = fread(file.path(logratios_dir, "Candidate_Orthologs_clr.tsv"), 
                                  data.table = FALSE)

## proportionality matrix
Col_ortho_rho = fread(file.path(logratios_dir, "Candidate_Orthologs_rho.tsv"), 
                                  data.table = FALSE)
# GOMixer annots
Col_ortho_GMM = fread(file.path(base_dir,
                                "code/metagenome_profiles/R_notebooks/Tables/Candidate_Orthologs_GMM.tsv"), 
                                    data.table = FALSE)

```

# Linear models with Maaslin2

I will use the Maaslin2 package to run linear models on the functional profile data while accounting for relevant co-variates.

## Write Maaslin2 wrapper function

Maaslin2 is a bit awkward to run, so I will create a wrapper function to make it work a bit better. Note that these are issues with version 1.0.0.

* It only works when the input data is written on a file on disk, not directly from R. Because of that, I will make the wrapper function write the input tables to temporary files and executin the function. [See this issue](https://github.com/biobakery/Maaslin2/issues/1).

* The filtering doesn't really work, so I have to manually filter the tables and then use those filtered as input. [See this issue](https://forum.biobakery.org/t/filtering-of-features-by-abundance/301)

* The function is very verbose and there is no command to deactivate it, so I'm sending the output to a temporary log file.

* Some of the presets are better turned off, so I will configure the defaults as I see fit.

```{r}
# Define function with new defaults
maaslin_2 = function(input_data = input_data,
                     input_metadata = input_metadata,
                     output = "/tmp/global2/jdelacuesta/R",
                     temp_dir = "/tmp/global2/jdelacuesta/R",
                     qval = 0.1,
                     min_abundance = -1,
                     min_prevalence = 0,
                     normalization = "NONE",
                     transform = "NONE",
                     analysis_method = "LM",
                     max_significance = 0.25,
                     random_effects = NULL,
                     fixed_effects = NULL,
                     correction = "BH",
                     standardize = FALSE,
                     cores = 1,
                     plot_heatmap = FALSE,
                     plot_scatter = FALSE,
                     heatmap_first_n = 50) {
  
  # Write data frames to temporary files to execute Maaslin2
  tmp_feature = tempfile(pattern = "feats_", tmpdir = "/ebio/abt3_projects/small_projects/jdelacuesta/tmp", fileext = ".txt")
  write_delim(input_data, tmp_feature, delim = "\t")

  tmp_metadata = tempfile(pattern = "metadata_", tmpdir = "/ebio/abt3_projects/small_projects/jdelacuesta/tmp", fileext = ".txt")
  write_delim(input_metadata, tmp_metadata, delim = "\t")
  
  # Maaslin is very verbose
  # There is no way of deactivating the log
  # I'll sink it to a temporary file as well
  tmplog = tempfile(pattern = "log_", tmpdir = "/ebio/abt3_projects/small_projects/jdelacuesta/tmp", fileext = ".txt")
  #print(tmplog)
  sink(tmplog)
  
  # Execute Maaslin2
  maaslin_2_fit = Maaslin2::Maaslin2(input_data = tmp_feature, 
                     input_metadata = tmp_metadata, 
                     output = output, 
                     min_abundance = min_abundance, 
                     min_prevalence = min_prevalence, 
                     normalization = normalization, 
                     transform = transform, 
                     analysis_method = analysis_method, 
                     max_significance = max_significance, 
                     random_effects = random_effects, 
                     fixed_effects = fixed_effects, 
                     correction = correction, 
                     standardize = standardize, 
                     cores = cores, 
                     plot_heatmap = plot_heatmap, 
                     plot_scatter = plot_scatter, 
                     heatmap_first_n = heatmap_first_n)
  # End log sink
  sink() 
  
  # Filter the results table by qvalue threshold
  maaslin_2_df = maaslin_2_fit$result
  maaslin_2_df
}

# Function to easily run the wrapper and extract table with values
run_maaslin_2 = function(Features, Metadata, Covariates, Explanatory, padj = 0.1, full_results = FALSE) {
  # Execute maaslin_2 wrapper function
  maaslin_result = maaslin_2(Features, 
                             Metadata, 
                             fixed_effects = c(Covariates, Explanatory))
  if(full_results == TRUE){
    maaslin_result %>% 
  filter(qval <= padj)
  } else {
      # Filter result
  maaslin_result %>% 
  filter(qval <= padj, metadata == Explanatory)
  }
}
```

## Prepare data
### Create data frame of metadata
The target variable and the covariates should be in the same data frame.
```{r}
# Dataframe with appropriate covariates
Col_metadata_full = Col_metadata %>% 
  select(ID, City, Sex, Age,
         BMI, BMI_status,
         Waist_Circumference, Waist_Circumference_status, 
         Fat_percentage, Fat_percentage_status,
         bsp_class, chs_class, MS_score, ends_with("_med"))

Col_metadata_full %>% dfhead()

# Data frame for sensitivity analysis
Col_metadata_sens = Col_metadata_full %>% 
  filter(n_med == 0)
```

## Run Maaslin2 on KEGG modules
```{r}
# Definition of modules
module_annot = Col_module_summ %>% 
  select(Module, Definition, g.mean, msd, Prevalence)
```

### No adjustment by CHS
*Co-variates to adjust*: I will fit the linear models using design-controlled variables as covariates. 
```{r}
# Create a vector of the covariates I will use to adjust
# The test variable will be added during the function call
covariates_vector_ind = c("City", "Age", "Sex", "Hypertension_med", "Diabetes_med", "Dislipid_med", "PPI_med")
```

*Explanatory variables to test*
```{r}
explanatory_vector_ind = c("Fat_percentage_status", "Fat_percentage", "BMI", 
                      "BMI_status", "Waist_Circumference", "Waist_Circumference_status",
                      "bsp_class", "chs_class", "MS_score")
```

```{r}
# Run Maaslin2 on all the explanatory variables and save in list
full_all_noadj = map(explanatory_vector_ind, 
    function(expl) run_maaslin_2(Col_module_clr, Col_metadata_full, covariates_vector_ind, expl))

names(full_all_noadj) = explanatory_vector_ind
```

### Adjusting by CHS

*Co-variates to adjust*: In addition to the already used variables, I will include the CHS as covariate to adjust the models
```{r}
# Create a vector of the covariates I will use to adjust
# The test variable will be added during the function call
covariates_vector_adj = c("City", "Age", "Sex", "Hypertension_med", "Diabetes_med", "Dislipid_med", "PPI_med", "chs_class")
```

*Explanatory variables to test* remove the CHS status from the vector of variables since it is now included in the covariate vector.
```{r}
explanatory_vector_adj = c("Fat_percentage_status", "Fat_percentage", "BMI", 
                           "BMI_status", "Waist_Circumference", "Waist_Circumference_status",
                           "bsp_class", "MS_score")
```

```{r}
# Run Maaslin2 on all the explanatory variables and save in list
full_all_adj = map(explanatory_vector_adj, 
    function(expl) run_maaslin_2(Col_module_clr, Col_metadata_full, covariates_vector_adj, expl, full_results = TRUE))

names(full_all_adj) = explanatory_vector_adj
```


### BMI as explanatory variable
Shared features are those whose association with a given host parameter disappears once adjusted for a second host parameter; shared features do not necessarily have significant associations with both host parameters. On the other hand, unique features are those whose association with a parameter remains even after accounting for the second one.
```{r}
# Distinct significant modules associated with BMI
BMI_mod = full_all_noadj %>% 
  pluck("BMI") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant modules associated with CHS
CHS_mod = full_all_noadj %>% 
  pluck("chs_class") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant modules associated with BMI adj for CHS
# i.e. unique
adj_BMI_mod = full_all_adj %>% 
  pluck("BMI") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "BMI") %>% 
  mutate(metadata = str_c(metadata, "_adjCHS"))

# Distinct significant modules associated with CHS adj for BMI
# i.e. unique
adj_bmiCHS_mod = full_all_adj %>% 
  pluck("BMI") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "chs_class") %>% 
  mutate(metadata = str_c(metadata, "_adjBMI"))

# All modules associated with CHS and BMI. Tangled association
pool_BMI_CHS_mod = bind_rows(BMI_mod, CHS_mod, adj_BMI_mod, adj_bmiCHS_mod)

# Core associations
# Modules associated with BMI confounded by CHS
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_BMI_mod = bind_rows(anti_join(BMI_mod, adj_BMI_mod, by = c("feature")), 
                             anti_join(adj_BMI_mod, BMI_mod, by = c("feature"))) %>% 
  arrange(feature)

# Pathways associated with CHS confounded by BMI
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_CHS_mod = bind_rows(anti_join(CHS_mod, adj_bmiCHS_mod, by = c("feature")), 
                             anti_join(adj_bmiCHS_mod, CHS_mod, by = c("feature"))) %>% 
  arrange(feature)

# These anti_joins are features that are present in models with a single main effect but absent in models with two main effects
# Think of each confound_* table as the core contributed by each main effect
# Therefore they are not part of the unique, because it was established they are mediated by one of the main effects

# All pathways confounded by BMI or CHS
core_BMI_CHS_mod = pool_BMI_CHS_mod %>%
  filter((feature %in% c(confound_BMI_mod$feature, confound_CHS_mod$feature))) %>% 
  arrange(feature) %>% 
  add_count(feature)


# Unique associations
# i.e associations after adjusting but not present in confounded set
uniq_BMI_mod = adj_BMI_mod %>% 
  filter(!(feature %in% core_BMI_CHS_mod$feature))

uniq_bmiCHS_mod = adj_bmiCHS_mod %>% 
  filter(!(feature %in% core_BMI_CHS_mod$feature))
```

Make sure that those that appear more than once have the same direction
```{r}
# positive BMI coef sign -> positive corr with BMI
# Positive CHS coef sign -> Enriched in healthy subjects/depleted in unhealthy
core_BMI_CHS_mod %>% 
  arrange(n, feature)

# n of 1 means that is associated with a single manin efect and lost when adjusted
# n of 2 means that is associated with the two single main effects but not when adjusted
# n of 3 mean that is associated with the two main effects and also when adjusting
```

Single table of significant modules count
```{r}
# Number of modules in each of the categories
# List of result tables
BMI_CHS_modules_list = list(pool_BMI_CHS_mod, BMI_mod, 
                            CHS_mod, adj_BMI_mod, 
                            adj_bmiCHS_mod, core_BMI_CHS_mod, 
                            uniq_BMI_mod, uniq_bmiCHS_mod)

names(BMI_CHS_modules_list) = c("pool_BMI_CHS_mod", "BMI_mod",
                                "CHS_mod", "adj_BMI_mod",
                                "adj_bmiCHS_mod", "core_BMI_CHS_mod", 
                                "unique_BMI_mod", "unique_CHS_mod")

# Create single table with numbers
BMI_CHS_modules_n = map(BMI_CHS_modules_list, function(x){
  x %>% 
    select(feature) %>% 
    distinct() %>% 
    nrow()
})

BMI_CHS_modules_n = BMI_CHS_modules_n %>% 
  unlist %>% 
  data.frame(Set = names(BMI_CHS_modules_list), n_modules = .)

BMI_CHS_modules_n
```

Add variable of health association to core and unique tables
```{r}
# Tables with annotation of unique and shared modules for each condition
uniq_BMI_mod_annot = uniq_BMI_mod %>% 
  left_join(module_annot, by = c("feature" = "Module")) %>% 
  mutate(quick_assoc = case_when(value == "BMI" & coef > 0 ~ "Disease", 
                                 value == "BMI" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Definition, quick_assoc, everything()) %>% 
  arrange(quick_assoc)


uniq_bmiCHS_mod_annot = uniq_bmiCHS_mod %>% 
  left_join(module_annot, by = c("feature" = "Module")) %>% 
  mutate(quick_assoc = case_when(value == "BMI" & coef > 0 ~ "Disease", 
                                 value == "BMI" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Definition, quick_assoc, everything()) %>% 
  arrange(quick_assoc)

BMI_CHS_mod_annot = core_BMI_CHS_mod %>% 
  left_join(module_annot, by = c("feature" = "Module")) %>% 
  mutate(quick_assoc = case_when(value == "BMI" & coef > 0 ~ "Disease", 
                                 value == "BMI" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Definition, quick_assoc, everything()) %>% 
  arrange(quick_assoc)
```

```{r}
# Wrirte tables
# write_tsv(uniq_BMI_mod_annot, file.path(tab_dir, "uniq_BMI_mod_annot.tsv"))
# write_tsv(uniq_bmiCHS_mod_annot, file.path(tab_dir, "uniq_bmiCHS_mod_annot.tsv"))
# write_tsv(BMI_CHS_mod_annot, file.path(tab_dir, "BMI_CHS_mod_annot.tsv"))
```

Summarize the abundance, prevalence and dispersion of the significant modules
```{r}
# Core modules
BMI_CHS_mod_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for CHS
uniq_bmiCHS_mod_annot %>%
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for BMI
uniq_BMI_mod_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))
```


### Waist circumference as explanatory variable
```{r}
# Distinct significant modules associated with WC
WC_mod = full_all_noadj %>% 
  pluck("Waist_Circumference") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant modules associated with CHS
CHS_mod = full_all_noadj %>% 
  pluck("chs_class") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant modules associated with WC adj for CHS
# i.e. unique
adj_WC_mod = full_all_adj %>% 
  pluck("Waist_Circumference") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "Waist_Circumference") %>% 
  mutate(metadata = str_c(metadata, "_adjCHS"))

# Distinct significant modules associated with CHS adj for WC
# i.e. unique
adj_wcCHS_mod = full_all_adj %>% 
  pluck("Waist_Circumference") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "chs_class") %>% 
  mutate(metadata = str_c(metadata, "_adjWC"))

# All modules associated with CHS and WC Tangled association
pool_WC_CHS_mod = bind_rows(WC_mod, CHS_mod, adj_WC_mod, adj_wcCHS_mod)

# Core associations
# Modules associated with WC confounded by CHS
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_WC_mod = bind_rows(anti_join(WC_mod, adj_WC_mod, by = c("feature")), 
                             anti_join(adj_WC_mod, WC_mod, by = c("feature"))) %>% 
  arrange(feature)

# Pathways associated with CHS confounded by WC
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_CHS_mod = bind_rows(anti_join(CHS_mod, adj_wcCHS_mod, by = c("feature")), 
                             anti_join(adj_wcCHS_mod, CHS_mod, by = c("feature"))) %>% 
  arrange(feature)

# All pathways confounded by WC or CHS
core_WC_CHS_mod = pool_WC_CHS_mod %>%
  filter((feature %in% c(confound_WC_mod$feature, confound_CHS_mod$feature))) %>% 
  arrange(feature) %>% 
  add_count(feature)

# Unique associations
# i.e associations after adjusting but not present in confounded set
uniq_WC_mod = adj_WC_mod %>% 
  filter(!(feature %in% core_WC_CHS_mod$feature))

uniq_wcCHS_mod = adj_wcCHS_mod %>% 
  filter(!(feature %in% core_WC_CHS_mod$feature))
```

Make sure that those that appear more than once have the same direction
```{r}
# positive WC coef sign -> positive corr with WC
# Positive CHS coef sign -> Enriched in healthy subjects/depleted in unhealthy
core_WC_CHS_mod %>% 
  arrange(n, feature)

# n of 1 means that is associated with a single manin efect and lost when adjusted
# n of 2 means that is associated with the two single main effects but not when adjusted
# n of 3 mean that is associated with the two main effects and also when adjusting
```

Single table of significant modules count
```{r}
# Number of modules in each of the categories
# List of result tables
WC_CHS_modules_list = list(pool_WC_CHS_mod, WC_mod, 
                            CHS_mod, adj_WC_mod, 
                            adj_wcCHS_mod, core_WC_CHS_mod, 
                            uniq_WC_mod, uniq_wcCHS_mod)

names(WC_CHS_modules_list) = c("pool_WC_CHS_mod", "WC_mod",
                                "CHS_mod", "adj_WC_mod",
                                "adj_wcCHS_mod", "core_WC_CHS_mod", 
                                "unique_WC_mod", "unique_CHS_mod")

# Create single table with numbers
WC_CHS_modules_n = map(WC_CHS_modules_list, function(x){
  x %>% 
    select(feature) %>% 
    distinct() %>% 
    nrow()
})

WC_CHS_modules_n = WC_CHS_modules_n %>% 
  unlist %>% 
  data.frame(Set = names(WC_CHS_modules_list), n_modules = .)

WC_CHS_modules_n
```

Add variable of health association to core and unique tables
```{r}
# Tables with annotation of unique and shared modules for each condition
uniq_WC_mod_annot = uniq_WC_mod %>% 
  left_join(module_annot, by = c("feature" = "Module")) %>% 
  mutate(quick_assoc = case_when(value == "Waist_Circumference" & coef > 0 ~ "Disease", 
                                 value == "Waist_Circumference" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Definition, quick_assoc, everything()) %>% 
  arrange(quick_assoc)


uniq_wcCHS_mod_annot = uniq_wcCHS_mod %>% 
  left_join(module_annot, by = c("feature" = "Module")) %>% 
  mutate(quick_assoc = case_when(value == "Waist_Circumference" & coef > 0 ~ "Disease", 
                                 value == "Waist_Circumference" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Definition, quick_assoc, everything()) %>% 
  arrange(quick_assoc)

WC_CHS_mod_annot = core_WC_CHS_mod %>% 
  left_join(module_annot, by = c("feature" = "Module")) %>% 
  mutate(quick_assoc = case_when(value == "Waist_Circumference" & coef > 0 ~ "Disease", 
                                 value == "Waist_Circumference" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Definition, quick_assoc, everything()) %>% 
  arrange(quick_assoc)
```

```{r}
# Wrirte tables
# write_tsv(uniq_WC_mod_annot, file.path(tab_dir, "uniq_WC_mod_annot.tsv"))
# write_tsv(uniq_wcCHS_mod_annot, file.path(tab_dir, "uniq_wcCHS_mod_annot.tsv"))
# write_tsv(WC_CHS_mod_annot, file.path(tab_dir, "WC_CHS_mod_annot.tsv"))
```

Summarize the abundance, prevalence and dispersion of the significant modules
```{r}
# Core modules
WC_CHS_mod_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for CHS
uniq_wcCHS_mod_annot %>%
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for BMI
uniq_WC_mod_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))
```


### Percentage body fat as explanatory variable
```{r}
# Distinct significant modules associated with BF
BF_mod = full_all_noadj %>% 
  pluck("Fat_percentage") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant modules associated with CHS
CHS_mod = full_all_noadj %>% 
  pluck("chs_class") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant modules associated with BF adj for CHS
# i.e. unique
adj_BF_mod = full_all_adj %>% 
  pluck("Fat_percentage") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "Fat_percentage") %>% 
  mutate(metadata = str_c(metadata, "_adjCHS"))

# Distinct significant modules associated with CHS adj for BF
# i.e. unique
adj_wcCHS_mod = full_all_adj %>% 
  pluck("Fat_percentage") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "chs_class") %>% 
  mutate(metadata = str_c(metadata, "_adjBF"))

# All modules associated with CHS and BF. Tangled association
pool_BF_CHS_mod = bind_rows(BF_mod, CHS_mod, adj_BF_mod, adj_wcCHS_mod)

# Core associations
# Modules associated with BF confounded by CHS
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_BF_mod = bind_rows(anti_join(BF_mod, adj_BF_mod, by = c("feature")), 
                             anti_join(adj_BF_mod, BF_mod, by = c("feature"))) %>% 
  arrange(feature)

# Pathways associated with CHS confounded by BF
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_CHS_mod = bind_rows(anti_join(CHS_mod, adj_wcCHS_mod, by = c("feature")), 
                             anti_join(adj_wcCHS_mod, CHS_mod, by = c("feature"))) %>% 
  arrange(feature)

# All pathways confounded by BF or CHS
core_BF_CHS_mod = pool_BF_CHS_mod %>%
  filter((feature %in% c(confound_BF_mod$feature, confound_CHS_mod$feature))) %>% 
  arrange(feature) %>% 
  add_count(feature)


# Unique associations
# i.e associations after adjusting but not present in confounded set
uniq_BF_mod = adj_BF_mod %>% 
  filter(!(feature %in% core_BF_CHS_mod$feature))

uniq_bfCHS_mod = adj_wcCHS_mod %>% 
  filter(!(feature %in% core_BF_CHS_mod$feature))
```

Make sure that those that appear more than once have the same direction
```{r}
# positive BF coef sign -> positive corr with BF
# Positive CHS coef sign -> Enriched in healthy subjects/depleted in unhealthy
core_BF_CHS_mod %>% 
  arrange(n, feature)

# n of 1 means that is associated with a single manin efect and lost when adjusted
# n of 2 means that is associated with the two single main effects but not when adjusted
# n of 3 mean that is associated with the two main effects and also when adjusting
```

Single table of significant modules count
```{r}
# Number of modules in each of the categories
# List of result tables
BF_CHS_modules_list = list(pool_BF_CHS_mod, BF_mod, 
                            CHS_mod, adj_BF_mod, 
                            adj_wcCHS_mod, core_BF_CHS_mod, 
                            uniq_BF_mod, uniq_bfCHS_mod)

names(BF_CHS_modules_list) = c("pool_BF_CHS_mod", "BF_mod",
                                "CHS_mod", "adj_BF_mod",
                                "adj_wcCHS_mod", "core_BF_CHS_mod", 
                                "unique_BF_mod", "unique_CHS_mod")

# Create single table with numbers
BF_CHS_modules_n = map(BF_CHS_modules_list, function(x){
  x %>% 
    select(feature) %>% 
    distinct() %>% 
    nrow()
})

BF_CHS_modules_n = BF_CHS_modules_n %>% 
  unlist %>% 
  data.frame(Set = names(BF_CHS_modules_list), n_modules = .)

BF_CHS_modules_n
```

Add variable of health association to core and unique tables
```{r}
# Tables with annotation of unique and shared modules for each condition
uniq_BF_mod_annot = uniq_BF_mod %>% 
  left_join(module_annot, by = c("feature" = "Module")) %>% 
  mutate(quick_assoc = case_when(value == "Fat_percentage" & coef > 0 ~ "Disease", 
                                 value == "Fat_percentage" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Definition, quick_assoc, everything()) %>% 
  arrange(quick_assoc)


uniq_bfCHS_mod_annot = uniq_bfCHS_mod %>% 
  left_join(module_annot, by = c("feature" = "Module")) %>% 
  mutate(quick_assoc = case_when(value == "Fat_percentage" & coef > 0 ~ "Disease", 
                                 value == "Fat_percentage" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Definition, quick_assoc, everything()) %>% 
  arrange(quick_assoc)

BF_CHS_mod_annot = core_BF_CHS_mod %>% 
  left_join(module_annot, by = c("feature" = "Module")) %>% 
  mutate(quick_assoc = case_when(value == "Fat_percentage" & coef > 0 ~ "Disease", 
                                 value == "Fat_percentage" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Definition, quick_assoc, everything()) %>% 
  arrange(quick_assoc)
```


Summarize the abundance, prevalence and dispersion of the significant modules
```{r}
# Core modules
BF_CHS_mod_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for CHS
uniq_bfCHS_mod_annot %>%
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for BF
uniq_BF_mod_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))
```

## Combine results of obesity measures
```{r}
# Core associations
BMI_CHS_mod_annot
WC_CHS_mod_annot
BF_CHS_mod_annot

# Unique to each obesity measure adj by CHS
uniq_BMI_mod_annot
uniq_WC_mod_annot
uniq_BF_mod_annot


# Unique to CHS adj by each obesity measure
uniq_bmiCHS_mod_annot
uniq_wcCHS_mod_annot
uniq_bfCHS_mod_annot

```

# Combine Module results
```{r}
bind_core_mod = bind_rows(mutate(BMI_CHS_mod_annot, adj = "BMI"), 
                   mutate(WC_CHS_mod_annot, adj = "Waist_Circumference"),
                   mutate(BF_CHS_mod_annot, adj = "Fat_percentage")) %>% 
  #select(feature, annot, quick_assoc, metadata, adj, value) %>% 
  group_by(adj, feature) %>% 
  slice(1) %>% 
  ungroup()

bind_chs_uniq_mod = bind_rows(mutate(uniq_bmiCHS_mod_annot, adj = "BMI"), 
                   mutate(uniq_wcCHS_mod_annot, adj = "Waist_Circumference"),
                   mutate(uniq_bfCHS_mod_annot, adj = "Fat_percentage"))

bind_ob_uniq_mod = bind_rows(uniq_BMI_mod_annot, uniq_WC_mod_annot, uniq_BF_mod_annot) 

# Features unique for at least one OB measure but shared/core in others
extended_core_mod = bind_ob_uniq_mod %>% 
  filter(feature %in% bind_core_mod$feature) %>% 
  arrange(feature)

# Features unique to at least one OB measure and not shared/core in others
strict_uniq_mod = bind_ob_uniq_mod %>% 
  filter(!(feature %in% bind_core_mod$feature)) %>% 
  arrange(feature)

```

```{r}
bind_ob_uniq_mod %>% 
  add_count(feature) %>% 
  arrange(quick_assoc, feature, metadata)
```

```{r}
bind_chs_uniq_mod %>% 
  count(feature, sort = TRUE)

bind_chs_uniq_mod %>% 
  arrange(feature, adj)
```

```{r}
bind_core_mod %>% 
  count(feature, sort = TRUE)

bind_core_mod %>% 
  arrange(feature, metadata)
```

## Abundance and prevalence distribution of significant modules
Are the disease-associated modules more or less abundant than the health associated ones?
```{r}
# BMI
# Test mean abundance and prevalence
uniq_BMI_mod_annot %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))

wilcox.test(g.mean ~ factor(quick_assoc), data = uniq_BMI_mod_annot, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = uniq_BMI_mod_annot, exact = FALSE)

# WC
# Test mean abundance and prevalence
uniq_WC_mod_annot %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))
wilcox.test(g.mean ~ factor(quick_assoc), data = uniq_WC_mod_annot, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = uniq_WC_mod_annot, exact = FALSE)

# BF
# Test mean abundance and prevalence
uniq_BF_mod_annot %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))
wilcox.test(g.mean ~ factor(quick_assoc), data = uniq_BF_mod_annot, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = uniq_BF_mod_annot, exact = FALSE)
```


```{r}
# Core BMI
BMI_CHS_mod_distinct = BMI_CHS_mod_annot %>% 
  select(feature, annot, quick_assoc, g.mean, Prevalence) %>% 
  distinct()

BMI_CHS_mod_distinct %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))

wilcox.test(g.mean ~ factor(quick_assoc), data = BMI_CHS_mod_distinct, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = BMI_CHS_mod_distinct, exact = FALSE)

# Core WC
WC_CHS_mod_distinct = WC_CHS_mod_annot %>% 
  select(feature, annot, quick_assoc, g.mean, Prevalence) %>% 
  distinct()

WC_CHS_mod_distinct %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))

wilcox.test(g.mean ~ factor(quick_assoc), data = WC_CHS_mod_distinct, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = WC_CHS_mod_distinct, exact = FALSE)

# Core BF
BF_CHS_mod_distinct = BF_CHS_mod_annot %>% 
  select(feature, annot, quick_assoc, g.mean, Prevalence) %>% 
  distinct()

BF_CHS_mod_distinct %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))

wilcox.test(g.mean ~ factor(quick_assoc), data = BF_CHS_mod_distinct, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = BF_CHS_mod_distinct, exact = FALSE)
```

```{r fig.height=9, fig.width=8}
# Plot differences in abundance/prevalence
uniq_BMI_ab_bxplt = uniq_BMI_mod_annot %>% 
  ggplot(aes(x = quick_assoc, y = g.mean)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75, color="grey50") +
    stat_summary(fun.data=mean_sdl, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "mean(clr-transformed abundance)") 

uniq_BMI_pr_bxplt = uniq_BMI_mod_annot %>%
  mutate(Prevalence = Prevalence *100) %>% 
  ggplot(aes(x = quick_assoc, y = Prevalence)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75, color="grey50") +
    scale_y_continuous(limits = c(97, 100), oob = scales::squish) +
    stat_summary(fun.data=mean_sdl, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "Prevalence (%)") 

BMI_CHS_ab_bxplt = BMI_CHS_mod_annot %>% 
  ggplot(aes(x = quick_assoc, y = g.mean)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75, color="grey50") +
    stat_summary(fun.data=mean_sdl, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "mean(clr-transformed abundance)") 

BMI_CHS_pr_bxplt = BMI_CHS_mod_annot %>% 
  mutate(Prevalence = Prevalence *100) %>% 
  ggplot(aes(x = quick_assoc, y = Prevalence)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75, color="grey50") +
    scale_y_continuous(limits = c(97, 100), oob = scales::squish) +
    stat_summary(fun.data=mean_sdl, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "Prevalence (%)") 

uniq_BMI_bxplts = plot_grid(uniq_BMI_ab_bxplt, uniq_BMI_pr_bxplt, align = "h", nrow = 1)
BMI_CHS_bxplts = plot_grid(BMI_CHS_ab_bxplt, BMI_CHS_pr_bxplt, align = "h", nrow = 1)
BMI_pr_ab_bxplots = plot_grid(BMI_CHS_bxplts, uniq_BMI_bxplts, align = "hv", nrow = 2, labels = "AUTO")
BMI_pr_ab_bxplots
save_plot(file.path(fig_dir, "BMI_pr_ab_bxplots.png"), BMI_pr_ab_bxplots,
          base_width = 8, base_height = 9)

```

```{r}
assoc_colors = c("Disease" = "#671646", "Health" = "#a8adc4")
sig_mods_boxplot_df = bind_rows(bind_ob_uniq_mod, bind_core_mod) %>% 
  bind_rows(bind_chs_uniq_mod) %>% 
  select(feature, quick_assoc, g.mean, Prevalence) %>% 
  distinct() 

mod_abund_plot = sig_mods_boxplot_df %>% 
  ggplot(aes(x = quick_assoc, y = g.mean, color = quick_assoc)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75) +
    scale_color_manual(values = assoc_colors) +
    stat_summary(fun.data=mean_se, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "mean(clr-transformed abundance)") +
    theme(legend.position = "none")

mod_prev_plot = sig_mods_boxplot_df %>%
  mutate(Prevalence = Prevalence *100) %>% 
  ggplot(aes(x = quick_assoc, y = Prevalence,  color = quick_assoc)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75) + 
    scale_color_manual(values = assoc_colors) +
    scale_y_continuous(limits = c(97, 100), oob = scales::squish) +
    stat_summary(fun.data=mean_se, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "Prevalence (%)") +
    theme(legend.position = "none")

sig_mods_boxplot_df %>% 
  rstatix::wilcox_test(Prevalence ~ quick_assoc)

sig_mods_boxplot_df %>% 
  rstatix::wilcox_test(g.mean ~ quick_assoc)

ggpubr::ggarrange(mod_abund_plot, mod_prev_plot)
```

```{r}
sig_ortho_boxplot_df = bind_rows(bind_ob_uniq_ortho, bind_core_ortho) %>% 
  bind_rows(bind_chs_uniq_ortho) %>% 
  select(feature, quick_assoc, g.mean, Prevalence) %>% 
  distinct() 

ortho_abund_plot = sig_ortho_boxplot_df %>% 
  ggplot(aes(x = quick_assoc, y = g.mean, color = quick_assoc)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.3) +
    scale_color_manual(values = assoc_colors) +
    stat_summary(fun.data=mean_se, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "mean(clr-transformed abundance)") +
    theme(legend.position = "none")

ortho_prev_plot = sig_ortho_boxplot_df %>%
  mutate(Prevalence = Prevalence *100) %>% 
  ggplot(aes(x = quick_assoc, y = Prevalence,  color = quick_assoc)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.3) + 
    scale_color_manual(values = assoc_colors) +
    scale_y_continuous(limits = c(45, 100), oob = scales::squish) +
    stat_summary(fun.data=mean_se, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "Prevalence (%)") +
    theme(legend.position = "none")

sig_ortho_boxplot_df %>% 
  rstatix::wilcox_test(Prevalence ~ quick_assoc)

sig_ortho_boxplot_df %>% 
  rstatix::wilcox_test(g.mean ~ quick_assoc)

ggpubr::ggarrange(ortho_abund_plot, ortho_prev_plot)

```

```{r fig.height=9, fig.width=7}
feature_prev_abund = ggpubr::ggarrange(mod_abund_plot, mod_prev_plot, ortho_abund_plot, ortho_prev_plot, labels = "AUTO")
save_plot(file.path(fig_dir, "Feature_pr_ab_bxplots.png"), feature_prev_abund,
          base_width = 8, base_height = 9)
```


## Write Module tables
```{r}
bind_ob_uniq_mod %>%
  write_tsv(., file.path(tab_dir, "ob_uniq_mod_meds.tsv"))

bind_chs_uniq_mod %>%
  write_tsv(., file.path(tab_dir, "chs_uniq_mod_meds.tsv"))

bind_core_mod %>%
  write_tsv(., file.path(tab_dir, "core_mod_meds.tsv"))

extended_core_mod %>% 
  write_tsv(., file.path(tab_dir, "extended_core_mod_meds.tsv"))

strict_uniq_mod %>% 
  write_tsv(., file.path(tab_dir, "strict_ob_uniq_mod_meds.tsv"))
```


# Run Maaslin2 on KEGG orthologs
```{r}
# Definition of orthologs
ortho_annot = Col_ortho_summ %>% 
  select(Ortholog, Description, g.mean, msd, Prevalence)
```

### No adjustment by CHS
*Co-variates to adjust*: I will fit the linear models using design-controlled variables as covariates. 
```{r}
# Create a vector of the covariates I will use to adjust
# The test variable will be added during the function call
covariates_vector_ind = c("City", "Age", "Sex", "Hypertension_med", "Diabetes_med", "Dislipid_med", "PPI_med")
```

*Explanatory variables to test*
```{r}
explanatory_vector_ind = c("Fat_percentage_status", "Fat_percentage", "BMI", 
                      "BMI_status", "Waist_Circumference", "Waist_Circumference_status",
                      "bsp_class", "chs_class", "MS_score")
```

```{r}
# Run Maaslin2 on all the explanatory variables and save in list
full_all_noadj = map(explanatory_vector_ind, 
    function(expl) run_maaslin_2(Col_ortho_clr, Col_metadata_full, covariates_vector_ind, expl))

names(full_all_noadj) = explanatory_vector_ind
```

### Adjusting by CHS

*Co-variates to adjust*: In addition to the already used variables, I will include the CHS as covariate to adjust the models
```{r}
# Create a vector of the covariates I will use to adjust
# The test variable will be added during the function call
covariates_vector_adj = c("City", "Age", "Sex", "Hypertension_med", "Diabetes_med", "Dislipid_med", "PPI_med", "chs_class")
```

*Explanatory variables to test* remove the CHS status from the vector of variables since it is now included in the covariate vector.
```{r}
explanatory_vector_adj = c("Fat_percentage_status", "Fat_percentage", "BMI", 
                           "BMI_status", "Waist_Circumference", "Waist_Circumference_status",
                           "bsp_class", "MS_score")
```

```{r}
# Run Maaslin2 on all the explanatory variables and save in list
full_all_adj = map(explanatory_vector_adj, 
    function(expl) run_maaslin_2(Col_ortho_clr, Col_metadata_full, covariates_vector_adj, expl, full_results = TRUE))

names(full_all_adj) = explanatory_vector_adj
```


### BMI as explanatory variable
Shared features are those whose association with a given host parameter disappears once adjusted for a second host parameter; shared features do not necessarily have significant associations with both host parameters. On the other hand, unique features are those whose association with a parameter remains even after accounting for the second one.
```{r}
# Distinct significant orthologs associated with BMI
BMI_ortho = full_all_noadj %>% 
  pluck("BMI") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant orthologs associated with CHS
CHS_ortho = full_all_noadj %>% 
  pluck("chs_class") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant orthologs associated with BMI adj for CHS
# i.e. unique
adj_BMI_ortho = full_all_adj %>% 
  pluck("BMI") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "BMI") %>% 
  mutate(metadata = str_c(metadata, "_adjCHS"))

# Distinct significant orthologs associated with CHS adj for BMI
# i.e. unique
adj_bmiCHS_ortho = full_all_adj %>% 
  pluck("BMI") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "chs_class") %>% 
  mutate(metadata = str_c(metadata, "_adjBMI"))

# All orthologs associated with CHS and BMI. Tangled association
pool_BMI_CHS_ortho = bind_rows(BMI_ortho, CHS_ortho, adj_BMI_ortho, adj_bmiCHS_ortho)

# Core associations
# Modules associated with BMI confounded by CHS
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_BMI_ortho = bind_rows(anti_join(BMI_ortho, adj_BMI_ortho, by = c("feature")), 
                             anti_join(adj_BMI_ortho, BMI_ortho, by = c("feature"))) %>% 
  arrange(feature)

# Pathways associated with CHS confounded by BMI
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_CHS_ortho = bind_rows(anti_join(CHS_ortho, adj_bmiCHS_ortho, by = c("feature")), 
                             anti_join(adj_bmiCHS_ortho, CHS_ortho, by = c("feature"))) %>% 
  arrange(feature)

# All pathways confounded by BMI or CHS
core_BMI_CHS_ortho = pool_BMI_CHS_ortho %>%
  filter((feature %in% c(confound_BMI_ortho$feature, confound_CHS_ortho$feature))) %>% 
  arrange(feature) %>% 
  add_count(feature)


# Unique associations
# i.e associations after adjusting but not present in confounded set
uniq_BMI_ortho = adj_BMI_ortho %>% 
  filter(!(feature %in% core_BMI_CHS_ortho$feature))

uniq_bmiCHS_ortho = adj_bmiCHS_ortho %>% 
  filter(!(feature %in% core_BMI_CHS_ortho$feature))
```

Make sure that those that appear more than once have the same direction
```{r}
# positive BMI coef sign -> positive corr with BMI
# Positive CHS coef sign -> Enriched in healthy subjects/depleted in unhealthy
core_BMI_CHS_ortho %>% 
  arrange(n, feature)

# n of 1 means that is associated with a single manin efect and lost when adjusted
# n of 2 means that is associated with the two single main effects but not when adjusted
# n of 3 mean that is associated with the two main effects and also when adjusting
```

Single table of significant orthologs count
```{r}
# Number of orthologs in each of the categories
# List of result tables
BMI_CHS_ortho_list = list(pool_BMI_CHS_ortho, BMI_ortho, 
                            CHS_ortho, adj_BMI_ortho, 
                            adj_bmiCHS_ortho, core_BMI_CHS_ortho, 
                            uniq_BMI_ortho, uniq_bmiCHS_ortho)

names(BMI_CHS_ortho_list) = c("pool_BMI_CHS_ortho", "BMI_ortho",
                                "CHS_ortho", "adj_BMI_ortho",
                                "adj_bmiCHS_ortho", "core_BMI_CHS_ortho", 
                                "unique_BMI_ortho", "unique_CHS_ortho")

# Create single table with numbers
BMI_CHS_ortho_n = map(BMI_CHS_ortho_list, function(x){
  x %>% 
    select(feature) %>% 
    distinct() %>% 
    nrow()
})

BMI_CHS_ortho_n = BMI_CHS_ortho_n %>% 
  unlist %>% 
  data.frame(Set = names(BMI_CHS_ortho_list), n_orthoules = .)

BMI_CHS_ortho_n
```

Add variable of health association to core and unique tables
```{r}
# Tables with annotation of unique and shared orthologs for each condition
uniq_BMI_ortho_annot = uniq_BMI_ortho %>% 
  left_join(ortho_annot, by = c("feature" = "Ortholog")) %>% 
  mutate(quick_assoc = case_when(value == "BMI" & coef > 0 ~ "Disease", 
                                 value == "BMI" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Description, quick_assoc, everything()) %>% 
  arrange(quick_assoc)


uniq_bmiCHS_ortho_annot = uniq_bmiCHS_ortho %>% 
  left_join(ortho_annot, by = c("feature" = "Ortholog")) %>% 
  mutate(quick_assoc = case_when(value == "BMI" & coef > 0 ~ "Disease", 
                                 value == "BMI" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Description, quick_assoc, everything()) %>% 
  arrange(quick_assoc)

BMI_CHS_ortho_annot = core_BMI_CHS_ortho %>% 
  left_join(ortho_annot, by = c("feature" = "Ortholog")) %>% 
  mutate(quick_assoc = case_when(value == "BMI" & coef > 0 ~ "Disease", 
                                 value == "BMI" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Description, quick_assoc, everything()) %>% 
  arrange(quick_assoc)
```

```{r}
# Wrirte tables
# write_tsv(uniq_BMI_ortho_annot, file.path(tab_dir, "uniq_BMI_ortho_annot.tsv"))
# write_tsv(uniq_bmiCHS_ortho_annot, file.path(tab_dir, "uniq_bmiCHS_ortho_annot.tsv"))
# write_tsv(BMI_CHS_ortho_annot, file.path(tab_dir, "BMI_CHS_ortho_annot.tsv"))
```

Summarize the abundance, prevalence and dispersion of the significant orthologs
```{r}
# Core orthologs
BMI_CHS_ortho_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for CHS
uniq_bmiCHS_ortho_annot %>%
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for BMI
uniq_BMI_ortho_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))
```


### Waist circumference as explanatory variable
```{r}
# Distinct significant orthologs associated with WC
WC_ortho = full_all_noadj %>% 
  pluck("Waist_Circumference") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant orthologs associated with CHS
CHS_ortho = full_all_noadj %>% 
  pluck("chs_class") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant orthologs associated with WC adj for CHS
# i.e. unique
adj_WC_ortho = full_all_adj %>% 
  pluck("Waist_Circumference") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "Waist_Circumference") %>% 
  mutate(metadata = str_c(metadata, "_adjCHS"))

# Distinct significant orthologs associated with CHS adj for WC
# i.e. unique
adj_wcCHS_ortho = full_all_adj %>% 
  pluck("Waist_Circumference") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "chs_class") %>% 
  mutate(metadata = str_c(metadata, "_adjWC"))

# All orthologs associated with CHS and WC Tangled association
pool_WC_CHS_ortho = bind_rows(WC_ortho, CHS_ortho, adj_WC_ortho, adj_wcCHS_ortho)

# Core associations
# Modules associated with WC confounded by CHS
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_WC_ortho = bind_rows(anti_join(WC_ortho, adj_WC_ortho, by = c("feature")), 
                             anti_join(adj_WC_ortho, WC_ortho, by = c("feature"))) %>% 
  arrange(feature)

# Pathways associated with CHS confounded by WC
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_CHS_ortho = bind_rows(anti_join(CHS_ortho, adj_wcCHS_ortho, by = c("feature")), 
                             anti_join(adj_wcCHS_ortho, CHS_ortho, by = c("feature"))) %>% 
  arrange(feature)

# All pathways confounded by WC or CHS
core_WC_CHS_ortho = pool_WC_CHS_ortho %>%
  filter((feature %in% c(confound_WC_ortho$feature, confound_CHS_ortho$feature))) %>% 
  arrange(feature) %>% 
  add_count(feature)

# Unique associations
# i.e associations after adjusting but not present in confounded set
uniq_WC_ortho = adj_WC_ortho %>% 
  filter(!(feature %in% core_WC_CHS_ortho$feature))

uniq_wcCHS_ortho = adj_wcCHS_ortho %>% 
  filter(!(feature %in% core_WC_CHS_ortho$feature))
```

Make sure that those that appear more than once have the same direction
```{r}
# positive WC coef sign -> positive corr with WC
# Positive CHS coef sign -> Enriched in healthy subjects/depleted in unhealthy
core_WC_CHS_ortho %>% 
  arrange(n, feature)

# n of 1 means that is associated with a single manin efect and lost when adjusted
# n of 2 means that is associated with the two single main effects but not when adjusted
# n of 3 mean that is associated with the two main effects and also when adjusting
```

Single table of significant orthologs count
```{r}
# Number of orthologs in each of the categories
# List of result tables
WC_CHS_ortho_list = list(pool_WC_CHS_ortho, WC_ortho, 
                            CHS_ortho, adj_WC_ortho, 
                            adj_wcCHS_ortho, core_WC_CHS_ortho, 
                            uniq_WC_ortho, uniq_wcCHS_ortho)

names(WC_CHS_ortho_list) = c("pool_WC_CHS_ortho", "WC_ortho",
                                "CHS_ortho", "adj_WC_ortho",
                                "adj_wcCHS_ortho", "core_WC_CHS_ortho", 
                                "unique_WC_ortho", "unique_CHS_ortho")

# Create single table with numbers
WC_CHS_ortho_n = map(WC_CHS_ortho_list, function(x){
  x %>% 
    select(feature) %>% 
    distinct() %>% 
    nrow()
})

WC_CHS_ortho_n = WC_CHS_ortho_n %>% 
  unlist %>% 
  data.frame(Set = names(WC_CHS_ortho_list), n_orthoules = .)

WC_CHS_ortho_n
```

Add variable of health association to core and unique tables
```{r}
# Tables with annotation of unique and shared orthologs for each condition
uniq_WC_ortho_annot = uniq_WC_ortho %>% 
  left_join(ortho_annot, by = c("feature" = "Ortholog")) %>% 
  mutate(quick_assoc = case_when(value == "Waist_Circumference" & coef > 0 ~ "Disease", 
                                 value == "Waist_Circumference" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Description, quick_assoc, everything()) %>% 
  arrange(quick_assoc)


uniq_wcCHS_ortho_annot = uniq_wcCHS_ortho %>% 
  left_join(ortho_annot, by = c("feature" = "Ortholog")) %>% 
  mutate(quick_assoc = case_when(value == "Waist_Circumference" & coef > 0 ~ "Disease", 
                                 value == "Waist_Circumference" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Description, quick_assoc, everything()) %>% 
  arrange(quick_assoc)

WC_CHS_ortho_annot = core_WC_CHS_ortho %>% 
  left_join(ortho_annot, by = c("feature" = "Ortholog")) %>% 
  mutate(quick_assoc = case_when(value == "Waist_Circumference" & coef > 0 ~ "Disease", 
                                 value == "Waist_Circumference" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Description, quick_assoc, everything()) %>% 
  arrange(quick_assoc)
```

```{r}
# Wrirte tables
# write_tsv(uniq_WC_ortho_annot, file.path(tab_dir, "uniq_WC_ortho_annot.tsv"))
# write_tsv(uniq_wcCHS_ortho_annot, file.path(tab_dir, "uniq_wcCHS_ortho_annot.tsv"))
# write_tsv(WC_CHS_ortho_annot, file.path(tab_dir, "WC_CHS_ortho_annot.tsv"))
```

Summarize the abundance, prevalence and dispersion of the significant orthologs
```{r}
# Core orthologs
WC_CHS_ortho_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for CHS
uniq_wcCHS_ortho_annot %>%
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for BMI
uniq_WC_ortho_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))
```


### Percentage body fat as explanatory variable
```{r}
# Distinct significant orthologs associated with BF
BF_ortho = full_all_noadj %>% 
  pluck("Fat_percentage") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant orthologs associated with CHS
CHS_ortho = full_all_noadj %>% 
  pluck("chs_class") %>% 
  select(feature, metadata, value, coef, pval, qval)

# Distinct significant orthologs associated with BF adj for CHS
# i.e. unique
adj_BF_ortho = full_all_adj %>% 
  pluck("Fat_percentage") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "Fat_percentage") %>% 
  mutate(metadata = str_c(metadata, "_adjCHS"))

# Distinct significant orthologs associated with CHS adj for BF
# i.e. unique
adj_wcCHS_ortho = full_all_adj %>% 
  pluck("Fat_percentage") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata == "chs_class") %>% 
  mutate(metadata = str_c(metadata, "_adjBF"))

# All orthologs associated with CHS and BF. Tangled association
pool_BF_CHS_ortho = bind_rows(BF_ortho, CHS_ortho, adj_BF_ortho, adj_wcCHS_ortho)

# Core associations
# Modules associated with BF confounded by CHS
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_BF_ortho = bind_rows(anti_join(BF_ortho, adj_BF_ortho, by = c("feature")), 
                             anti_join(adj_BF_ortho, BF_ortho, by = c("feature"))) %>% 
  arrange(feature)

# Pathways associated with CHS confounded by BF
# i.e present with no adjust, absent with adjust
# or present with adjust, absent with no adjust
confound_CHS_ortho = bind_rows(anti_join(CHS_ortho, adj_wcCHS_ortho, by = c("feature")), 
                             anti_join(adj_wcCHS_ortho, CHS_ortho, by = c("feature"))) %>% 
  arrange(feature)

# All pathways confounded by BF or CHS
core_BF_CHS_ortho = pool_BF_CHS_ortho %>%
  filter((feature %in% c(confound_BF_ortho$feature, confound_CHS_ortho$feature))) %>% 
  arrange(feature) %>% 
  add_count(feature)


# Unique associations
# i.e associations after adjusting but not present in confounded set
uniq_BF_ortho = adj_BF_ortho %>% 
  filter(!(feature %in% core_BF_CHS_ortho$feature))

uniq_bfCHS_ortho = adj_wcCHS_ortho %>% 
  filter(!(feature %in% core_BF_CHS_ortho$feature))
```

Make sure that those that appear more than once have the same direction
```{r}
# positive BF coef sign -> positive corr with BF
# Positive CHS coef sign -> Enriched in healthy subjects/depleted in unhealthy
core_BF_CHS_ortho %>% 
  arrange(n, feature)

# n of 1 means that is associated with a single manin efect and lost when adjusted
# n of 2 means that is associated with the two single main effects but not when adjusted
# n of 3 mean that is associated with the two main effects and also when adjusting
```

Single table of significant orthologs count
```{r}
# Number of orthologs in each of the categories
# List of result tables
BF_CHS_ortho_list = list(pool_BF_CHS_ortho, BF_ortho, 
                            CHS_ortho, adj_BF_ortho, 
                            adj_wcCHS_ortho, core_BF_CHS_ortho, 
                            uniq_BF_ortho, uniq_bfCHS_ortho)

names(BF_CHS_ortho_list) = c("pool_BF_CHS_ortho", "BF_ortho",
                                "CHS_ortho", "adj_BF_ortho",
                                "adj_wcCHS_ortho", "core_BF_CHS_ortho", 
                                "unique_BF_ortho", "unique_CHS_ortho")

# Create single table with numbers
BF_CHS_ortho_n = map(BF_CHS_ortho_list, function(x){
  x %>% 
    select(feature) %>% 
    distinct() %>% 
    nrow()
})

BF_CHS_ortho_n = BF_CHS_ortho_n %>% 
  unlist %>% 
  data.frame(Set = names(BF_CHS_ortho_list), n_orthoules = .)

BF_CHS_ortho_n
```

Add variable of health association to core and unique tables
```{r}
# Tables with annotation of unique and shared orthologs for each condition
uniq_BF_ortho_annot = uniq_BF_ortho %>% 
  left_join(ortho_annot, by = c("feature" = "Ortholog")) %>% 
  mutate(quick_assoc = case_when(value == "Fat_percentage" & coef > 0 ~ "Disease", 
                                 value == "Fat_percentage" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Description, quick_assoc, everything()) %>% 
  arrange(quick_assoc)


uniq_bfCHS_ortho_annot = uniq_bfCHS_ortho %>% 
  left_join(ortho_annot, by = c("feature" = "Ortholog")) %>% 
  mutate(quick_assoc = case_when(value == "Fat_percentage" & coef > 0 ~ "Disease", 
                                 value == "Fat_percentage" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Description, quick_assoc, everything()) %>% 
  arrange(quick_assoc)

BF_CHS_ortho_annot = core_BF_CHS_ortho %>% 
  left_join(ortho_annot, by = c("feature" = "Ortholog")) %>% 
  mutate(quick_assoc = case_when(value == "Fat_percentage" & coef > 0 ~ "Disease", 
                                 value == "Fat_percentage" & coef < 0 ~ "Health",
                                 value == "Healthy" & coef > 0 ~ "Health", 
                                 value == "Healthy" & coef < 0 ~ "Disease")) %>% 
  select(feature, annot = Description, quick_assoc, everything()) %>% 
  arrange(quick_assoc)
```

```{r}
# Wrirte tables
# write_tsv(uniq_BF_ortho_annot, file.path(tab_dir, "uniq_BF_ortho_annot.tsv"))
# write_tsv(uniq_wcCHS_ortho_annot, file.path(tab_dir, "uniq_wcCHS_ortho_annot.tsv"))
# write_tsv(BF_CHS_ortho_annot, file.path(tab_dir, "BF_CHS_ortho_annot.tsv"))
```

Summarize the abundance, prevalence and dispersion of the significant orthologs
```{r}
# Core orthologs
BF_CHS_ortho_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for CHS
uniq_bfCHS_ortho_annot %>%
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))

# Unique for BF
uniq_BF_ortho_annot %>% 
  select(feature, annot, quick_assoc, g.mean, msd, Prevalence) %>% 
  distinct() %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_mean = mean(g.mean), mean_sd = mean(msd), mean_prev = mean(Prevalence)) %>% 
  mutate_if(is.numeric, function(x) round(x, 3))
```

## Combine results of obesity measures
```{r}
# Core associations
BMI_CHS_ortho_annot
WC_CHS_ortho_annot
BF_CHS_ortho_annot

# Unique to each obesity measure adj by CHS
uniq_BMI_ortho_annot
uniq_WC_ortho_annot
uniq_BF_ortho_annot


# Unique to CHS adj by each obesity measure
uniq_bmiCHS_ortho_annot
uniq_wcCHS_ortho_annot
uniq_bfCHS_ortho_annot

```

# Combine Module results
```{r}
bind_core_ortho = bind_rows(mutate(BMI_CHS_ortho_annot, adj = "BMI"), 
                   mutate(WC_CHS_ortho_annot, adj = "Waist_Circumference"),
                   mutate(BF_CHS_ortho_annot, adj = "Fat_percentage")) %>% 
  #select(feature, annot, quick_assoc, metadata, adj, value) %>% 
  group_by(adj, feature) %>% 
  slice(1) %>% 
  ungroup()

bind_chs_uniq_ortho = bind_rows(mutate(uniq_bmiCHS_ortho_annot, adj = "BMI"), 
                   mutate(uniq_wcCHS_ortho_annot, adj = "Waist_Circumference"),
                   mutate(uniq_bfCHS_ortho_annot, adj = "Fat_percentage"))

bind_ob_uniq_ortho = bind_rows(uniq_BMI_ortho_annot, uniq_WC_ortho_annot, uniq_BF_ortho_annot) 

# Features unique for at least one OB measure but shared/core in others
extended_core_ortho = bind_ob_uniq_ortho %>% 
  filter(feature %in% bind_core_ortho$feature) %>% 
  arrange(feature)

# Features unique to at least one OB measure and not shared/core in others
strict_uniq_ortho = bind_ob_uniq_ortho %>% 
  filter(!(feature %in% bind_core_ortho$feature)) %>% 
  arrange(feature)

```

```{r}
bind_ob_uniq_ortho %>% 
  add_count(feature) %>% 
  arrange(quick_assoc, feature, metadata)
```

```{r}
bind_chs_uniq_ortho %>% 
  count(feature, sort = TRUE)

bind_chs_uniq_ortho %>% 
  arrange(feature, adj)
```

```{r}
bind_core_ortho %>% 
  count(feature, sort = TRUE)

bind_core_ortho %>% 
  arrange(feature, metadata)
```

## Abundance and prevalence distribution of significant orthologs
Are the disease-associated orthologs more or less abundant than the health associated ones?
```{r}
# BMI
# Test mean abundance and prevalence
uniq_BMI_ortho_annot %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))

wilcox.test(g.mean ~ factor(quick_assoc), data = uniq_BMI_ortho_annot, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = uniq_BMI_ortho_annot, exact = FALSE)

# WC
# Test mean abundance and prevalence
uniq_WC_ortho_annot %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))
wilcox.test(g.mean ~ factor(quick_assoc), data = uniq_WC_ortho_annot, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = uniq_WC_ortho_annot, exact = FALSE)

# BF
# Test mean abundance and prevalence
uniq_BF_ortho_annot %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))
wilcox.test(g.mean ~ factor(quick_assoc), data = uniq_BF_ortho_annot, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = uniq_BF_ortho_annot, exact = FALSE)
```


```{r}
# Core BMI
BMI_CHS_ortho_distinct = BMI_CHS_ortho_annot %>% 
  select(feature, annot, quick_assoc, g.mean, Prevalence) %>% 
  distinct()

BMI_CHS_ortho_distinct %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))

wilcox.test(g.mean ~ factor(quick_assoc), data = BMI_CHS_ortho_distinct, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = BMI_CHS_ortho_distinct, exact = FALSE)

# Core WC
WC_CHS_ortho_distinct = WC_CHS_ortho_annot %>% 
  select(feature, annot, quick_assoc, g.mean, Prevalence) %>% 
  distinct()

WC_CHS_ortho_distinct %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))

wilcox.test(g.mean ~ factor(quick_assoc), data = WC_CHS_ortho_distinct, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = WC_CHS_ortho_distinct, exact = FALSE)

# Core BF
BF_CHS_ortho_distinct = BF_CHS_ortho_annot %>% 
  select(feature, annot, quick_assoc, g.mean, Prevalence) %>% 
  distinct()

BF_CHS_ortho_distinct %>% 
  group_by(quick_assoc) %>% 
  summarise(mean_abund = mean(g.mean), mean_prev = mean(Prevalence))

wilcox.test(g.mean ~ factor(quick_assoc), data = BF_CHS_ortho_distinct, exact = FALSE)
wilcox.test(Prevalence ~ factor(quick_assoc), data = BF_CHS_ortho_distinct, exact = FALSE)
```

```{r fig.height=9, fig.width=8}
# Plot differences in abundance/prevalence
uniq_BMI_ab_bxplt = uniq_BMI_ortho_annot %>% 
  ggplot(aes(x = quick_assoc, y = g.mean)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75, color="grey50") +
    stat_summary(fun.data=mean_sdl, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "mean(clr-transformed abundance)") 

uniq_BMI_pr_bxplt = uniq_BMI_ortho_annot %>%
  mutate(Prevalence = Prevalence *100) %>% 
  ggplot(aes(x = quick_assoc, y = Prevalence)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75, color="grey50") +
    scale_y_continuous(limits = c(97, 100), oob = scales::squish) +
    stat_summary(fun.data=mean_sdl, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "Prevalence (%)") 

BMI_CHS_ab_bxplt = BMI_CHS_ortho_annot %>% 
  ggplot(aes(x = quick_assoc, y = g.mean)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75, color="grey50") +
    stat_summary(fun.data=mean_sdl, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "mean(clr-transformed abundance)") 

BMI_CHS_pr_bxplt = BMI_CHS_ortho_annot %>% 
  mutate(Prevalence = Prevalence *100) %>% 
  ggplot(aes(x = quick_assoc, y = Prevalence)) +
    geom_jitter(position=position_jitter(0.1), alpha = 0.75, color="grey50") +
    scale_y_continuous(limits = c(97, 100), oob = scales::squish) +
    stat_summary(fun.data=mean_sdl, fun.args = list(mult = 1), geom="pointrange", color="black", size = 0.75) +
    theme_light() +
    labs(x = "", y = "Prevalence (%)") 

uniq_BMI_bxplts = plot_grid(uniq_BMI_ab_bxplt, uniq_BMI_pr_bxplt, align = "h", nrow = 1)
BMI_CHS_bxplts = plot_grid(BMI_CHS_ab_bxplt, BMI_CHS_pr_bxplt, align = "h", nrow = 1)
BMI_pr_ab_bxplots = plot_grid(BMI_CHS_bxplts, uniq_BMI_bxplts, align = "hv", nrow = 2, labels = "AUTO")
BMI_pr_ab_bxplots
save_plot(file.path(fig_dir, "BMI_pr_ab_ortho_bxplots.png"), BMI_pr_ab_bxplots,
          base_width = 8, base_height = 9)

```

## Write Module tables
```{r}
bind_ob_uniq_ortho %>%
  write_tsv(., file.path(tab_dir, "ob_uniq_ortho_meds.tsv"))

bind_chs_uniq_ortho %>%
  write_tsv(., file.path(tab_dir, "chs_uniq_ortho_meds.tsv"))

bind_core_ortho %>%
  write_tsv(., file.path(tab_dir, "core_ortho_meds.tsv"))

extended_core_ortho %>% 
  write_tsv(., file.path(tab_dir, "extended_core_ortho_meds.tsv"))

strict_uniq_ortho %>% 
  write_tsv(., file.path(tab_dir, "strict_ob_uniq_ortho_meds.tsv"))
```

# LM coefficients plot
## Module
```{r}
maaslin_multiOB = map(explanatory_vector_adj, 
    function(expl) run_maaslin_2(Col_module_clr, Col_metadata_full, covariates_vector_adj, expl, full_results = TRUE, padj = 1))

names(maaslin_multiOB) = explanatory_vector_adj

Maaslin_OB_BMI = maaslin_multiOB %>% 
  pluck("BMI") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata %in% c("BMI", "chs_class"))

Maaslin_OB_WC = maaslin_multiOB %>% 
  pluck("Waist_Circumference") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata %in% c("Waist_Circumference", "chs_class"))

Maaslin_OB_BF = maaslin_multiOB %>% 
  pluck("Fat_percentage") %>% 
  select(feature, metadata, value, coef, pval, qval) %>% 
  filter(metadata %in% c("Fat_percentage", "chs_class"))
```

```{r}
# BMI
Maaslin_OB_BMI_wide = Maaslin_OB_BMI %>% 
  select(feature, metadata, coef, qval) %>% 
  pivot_wider(id_cols = feature, names_from = metadata, values_from = c(coef, qval)) %>% 
  mutate(feat_cat = case_when(feature %in% uniq_bmiCHS_mod_annot$feature ~ "CHS",
                               feature %in% uniq_BMI_mod_annot$feature ~ "BMI",
                               feature %in% BMI_CHS_mod_annot$feature ~ "Shared",
                               TRUE ~ "None"), 
         feat_cat = factor(feat_cat, levels = c("BMI", "CHS", "Shared", "None")))
Maaslin_OB_BMI_wide %>% dfhead()


# Waist circumference
Maaslin_OB_WC_wide = Maaslin_OB_WC %>% 
  select(feature, metadata, coef, qval) %>% 
  pivot_wider(id_cols = feature, names_from = metadata, values_from = c(coef, qval)) %>% 
  mutate(feat_cat = case_when(feature %in% uniq_wcCHS_mod_annot$feature ~ "CHS",
                               feature %in% uniq_WC_mod_annot$feature ~ "Waist Circumference",
                               feature %in% WC_CHS_mod_annot$feature ~ "Shared",
                               TRUE ~ "None"), 
         feat_cat = factor(feat_cat, levels = c("Waist Circumference", "CHS", "Shared", "None")))
Maaslin_OB_WC_wide %>% dfhead()

# BF
Maaslin_OB_BF_wide = Maaslin_OB_BF %>% 
  select(feature, metadata, coef, qval) %>% 
  pivot_wider(id_cols = feature, names_from = metadata, values_from = c(coef, qval)) %>% 
  mutate(feat_cat = case_when(feature %in% uniq_bfCHS_mod_annot$feature ~ "CHS",
                               feature %in% uniq_BF_mod_annot$feature ~ "Body Fat",
                               feature %in% BF_CHS_mod_annot$feature ~ "Shared",
                               TRUE ~ "None"), 
         feat_cat = factor(feat_cat, levels = c("Body Fat", "CHS", "Shared", "None")))

Maaslin_OB_BF_wide %>% dfhead()

```

```{r}
BMI_coef_plot = Maaslin_OB_BMI_wide %>%
  mutate(feat_cat = str_replace(feat_cat, "BMI", "OB")) %>% 
  ggplot(aes(x = coef_BMI, y = coef_chs_class)) +
    geom_hline(yintercept = 0, linetype="dashed") +
    geom_vline(xintercept = 0, linetype="dashed") +
    #geom_smooth(method = "lm", color = "grey30", se = FALSE) +
    geom_point(aes(color = feat_cat), alpha = 0.8) +
    scale_color_manual(values = c("OB" = "#0c2a85", "CHS" = "#F28E2B", "Shared" = "#76B7B2", "None" = "grey")) +
    theme_light() +
    labs(x = "BMI\nregression coefficient", 
         y = "CHS regression coefficient", 
         color = "Module\nassociation") +
    #theme(legend.position = "bottom", axis.title.y = element_blank()) +
    scale_y_continuous(expand = c(0, 0), limits = c(-0.5, 0.75)) +
    theme(text = element_text(size = 14))

WC_coef_plot = Maaslin_OB_WC_wide %>%
  mutate(feat_cat = str_replace(feat_cat, "Waist Circumference", "OB")) %>% 
  ggplot(aes(x = coef_Waist_Circumference, y = coef_chs_class)) +
    geom_hline(yintercept = 0, linetype="dashed") +
    geom_vline(xintercept = 0, linetype="dashed") +
    #geom_smooth(method = "lm", color = "grey30", se = FALSE) +
    geom_point(aes(color = feat_cat), alpha = 0.8) +
    scale_color_manual(values = c("OB" = "#0c2a85", "CHS" = "#F28E2B", "Shared" = "#76B7B2", "None" = "grey")) +
    theme_light() +
    #theme(legend.position = "bottom", axis.title.y = element_blank()) #+ scale_y_continuous(expand = c(0, 0), limits = c(-0.5, 0.75)) +
    labs(x = "WC\nregression coefficient", 
         y = "CHS regression coefficient", 
         color = "Module\nassociation") +
    theme(text = element_text(size = 14))


BF_coef_plot = Maaslin_OB_BF_wide %>%
  mutate(feat_cat = str_replace(feat_cat, "Body Fat", "OB")) %>% 
  ggplot(aes(x = coef_Fat_percentage, y = coef_chs_class)) +
    geom_hline(yintercept = 0, linetype="dashed") +
    geom_vline(xintercept = 0, linetype="dashed") +
    #geom_smooth(method = "lm", color = "grey30", se = FALSE) +
    geom_point(aes(color = feat_cat), alpha = 0.8) +
    scale_color_manual(values = c("OB" = "#0c2a85", "CHS" = "#F28E2B", "Shared" = "#76B7B2", "None" = "grey")) +
    theme_light() +
    labs(x = "BF%\nregression coefficient", 
         y = "CHS regression coefficient", 
         color = "Module\nassociation") +
    #theme(legend.position = "bottom", axis.title.y = element_blank()) + 
    scale_y_continuous(expand = c(0, 0), limits = c(-0.5, 0.8)) +
    scale_x_continuous(expand = c(0, 0), limits = c(-4, 6)) +
    theme(text = element_text(size = 14))
```

```{r fig.height=5, fig.width=12}
OB_coef_plot_hor = ggpubr::ggarrange(BMI_coef_plot, WC_coef_plot, BF_coef_plot, 
                  ncol = 3, 
                  align = "hv", 
                  common.legend = TRUE)

OB_coef_plot_hor

ggpubr::annotate_figure(OB_coef_plot_hor, left = "CHS regression coefficient") 

save_plot(file.path(fig_dir, "OB_coef_plot_hor.png"), OB_coef_plot_hor,
          base_height = 5, base_width = 12)
```

```{r}
# Very high CHS coefficient but not significant
large_nonsig = Maaslin_OB_BMI_wide %>% 
  filter(abs(coef_chs_class) > 0.3, feat_cat == "None") %>% 
  arrange(coef_chs_class)

large_nonsig


Col_module_summ %>% 
  filter(Module %in% large_nonsig$feature)
```


```{r fig.height=10, fig.width=4}
OB_coef_plot_vert = ggpubr::ggarrange(BMI_coef_plot, WC_coef_plot, BF_coef_plot, 
                  ncol = 1, 
                  align = "v", 
                  labels = c("A", "B", "C"), common.legend = TRUE)

OB_coef_plot_vert

save_plot(file.path(fig_dir, "OB_coef_plot_vert.png"), OB_coef_plot_vert,
          base_width = 3, base_height = 10)
```




```{r message=FALSE, warning=FALSE}
cor.test(Maaslin_OB_BMI_wide$coef_chs_class, Maaslin_OB_BMI_wide$coef_BMI, method = "spearman")
cor.test(Maaslin_OB_WC_wide$coef_chs_class, Maaslin_OB_WC_wide$coef_Waist_Circumference, method = "spearman")
cor.test(Maaslin_OB_BF_wide$coef_chs_class, Maaslin_OB_BF_wide$coef_Fat_percentage, method = "spearman")
```

```{r}
stop("Run plots with orthologs, rewrite some objects. If executing 'run all', continue manually")
```


# LM coefficients plot
## orthologs
```{r}
maaslin_multiOB = map(explanatory_vector_adj,
    function(expl) run_maaslin_2(Col_ortho_clr, Col_metadata_full, covariates_vector_adj, expl, full_results = TRUE, padj = 1))

names(maaslin_multiOB) = explanatory_vector_adj

Maaslin_OB_BMI = maaslin_multiOB %>%
  pluck("BMI") %>%
  select(feature, metadata, value, coef, pval, qval) %>%
  filter(metadata %in% c("BMI", "chs_class"))

Maaslin_OB_WC = maaslin_multiOB %>%
  pluck("Waist_Circumference") %>%
  select(feature, metadata, value, coef, pval, qval) %>%
  filter(metadata %in% c("Waist_Circumference", "chs_class"))

Maaslin_OB_BF = maaslin_multiOB %>%
  pluck("Fat_percentage") %>%
  select(feature, metadata, value, coef, pval, qval) %>%
  filter(metadata %in% c("Fat_percentage", "chs_class"))
```

```{r}
# BMI
Maaslin_OB_BMI_wide = Maaslin_OB_BMI %>% 
  select(feature, metadata, coef, qval) %>% 
  pivot_wider(id_cols = feature, names_from = metadata, values_from = c(coef, qval)) %>% 
  mutate(feat_cat = case_when(feature %in% uniq_bmiCHS_ortho_annot$feature ~ "CHS",
                               feature %in% uniq_BMI_ortho_annot$feature ~ "BMI",
                               feature %in% BMI_CHS_ortho_annot$feature ~ "Shared",
                               TRUE ~ "None"), 
         feat_cat = factor(feat_cat, levels = c("BMI", "CHS", "Shared", "None")))
Maaslin_OB_BMI_wide %>% dfhead()


# Waist circumference
Maaslin_OB_WC_wide = Maaslin_OB_WC %>% 
  select(feature, metadata, coef, qval) %>% 
  pivot_wider(id_cols = feature, names_from = metadata, values_from = c(coef, qval)) %>% 
  mutate(feat_cat = case_when(feature %in% uniq_wcCHS_ortho_annot$feature ~ "CHS",
                               feature %in% uniq_WC_ortho_annot$feature ~ "Waist Circumference",
                               feature %in% WC_CHS_ortho_annot$feature ~ "Shared",
                               TRUE ~ "None"), 
         feat_cat = factor(feat_cat, levels = c("Waist Circumference", "CHS", "Shared", "None")))
Maaslin_OB_WC_wide %>% dfhead()

# BF
Maaslin_OB_BF_wide = Maaslin_OB_BF %>% 
  select(feature, metadata, coef, qval) %>% 
  pivot_wider(id_cols = feature, names_from = metadata, values_from = c(coef, qval)) %>% 
  mutate(feat_cat = case_when(feature %in% uniq_bfCHS_ortho_annot$feature ~ "CHS",
                               feature %in% uniq_BF_ortho_annot$feature ~ "Body Fat",
                               feature %in% BF_CHS_ortho_annot$feature ~ "Shared",
                               TRUE ~ "None"), 
         feat_cat = factor(feat_cat, levels = c("Body Fat", "CHS", "Shared", "None")))

Maaslin_OB_BF_wide %>% dfhead()
```

```{r}
BMI_coef_plot = Maaslin_OB_BMI_wide %>%
  ggplot(aes(x = coef_BMI, y = coef_chs_class, color = feat_cat)) +
    geom_hline(yintercept = 0, linetype="dashed") +
    geom_vline(xintercept = 0, linetype="dashed") +
    #geom_smooth(method = "lm", color = "grey30", se = FALSE) +
    geom_point(data = subset(Maaslin_OB_BMI_wide, feat_cat %in% c("None")), alpha = 0.2) +
    geom_point(data = subset(Maaslin_OB_BMI_wide, feat_cat %in% c("BMI", "CHS", "Shared")), alpha = 0.8) +
    scale_color_manual(values = c("BMI" = "#0c2a85", "CHS" = "#F28E2B", "Shared" = "#76B7B2", "None" = "grey")) +
    theme_light() +
    labs(x = "BMI\nregression coefficient",
         y = "CHS regression coefficient",
         color = "Module\nAssociation") +
    theme(legend.position = "bottom",
          panel.border = element_rect(fill = NA, color = "lightgrey")) +
    scale_y_continuous(expand = c(0, 0), limits = c(-0.5, 0.75))

WC_coef_plot = Maaslin_OB_WC_wide %>%
  ggplot(aes(x = coef_Waist_Circumference, y = coef_chs_class, color = feat_cat)) +
    geom_hline(yintercept = 0, linetype="dashed") +
    geom_vline(xintercept = 0, linetype="dashed") +
    #geom_smooth(method = "lm", color = "grey30", se = FALSE) +
    geom_point(data = subset(Maaslin_OB_WC_wide, feat_cat %in% c("None")), alpha = 0.2) +
    geom_point(data = subset(Maaslin_OB_WC_wide, feat_cat %in% c("Waist Circumference", "CHS", "Shared")), alpha = 0.8) +
    scale_color_manual(values = c("Waist Circumference" = "#0c2a85", "CHS" = "#F28E2B", "Shared" = "#76B7B2", "None" = "grey")) +
    theme_light() +
    labs(x = "WC\nregression coefficient",
         y = "CHS regression coefficient",
         color = "Module\nAssociation") +
    theme(legend.position = "bottom",
          panel.border = element_rect(fill = NA, color = "lightgrey")) #+ scale_y_continuous(expand = c(0, 0), limits = c(-0.5, 0.75))

BF_coef_plot = Maaslin_OB_BF_wide %>%
  ggplot(aes(x = coef_Fat_percentage, y = coef_chs_class, color = feat_cat)) +
    geom_hline(yintercept = 0, linetype="dashed") +
    geom_vline(xintercept = 0, linetype="dashed") +
    #geom_smooth(method = "lm", color = "grey30", se = FALSE) +
    geom_point(data = subset(Maaslin_OB_BF_wide, feat_cat %in% c("None")), alpha = 0.2) +
    geom_point(data = subset(Maaslin_OB_BF_wide, feat_cat %in% c("Body Fat", "CHS", "Shared")), alpha = 0.8) +
    scale_color_manual(values = c("Body Fat" = "#0c2a85", "CHS" = "#F28E2B", "Shared" = "#76B7B2", "None" = "grey")) +
    theme_light() +
    labs(x = "BF%\nregression coefficient",
         y = "CHS regression coefficient",
         color = "Module\nAssociation") +
    theme(legend.position = "bottom",
          panel.border = element_rect(fill = NA, color = "lightgrey")) +
    scale_y_continuous(expand = c(0, 0), limits = c(-0.5, 0.8)) +
    scale_x_continuous(expand = c(0, 0), limits = c(-4, 6))
```

```{r fig.height=12, fig.width=5, message=FALSE, warning=FALSE}
OB_coef_plot_vert = ggpubr::ggarrange(BMI_coef_plot, WC_coef_plot, BF_coef_plot,
                  ncol = 1,
                  align = "hv", labels = c("A", "B", "C"))

OB_coef_plot_vert

save_plot(file.path(fig_dir, "OB_coef_plot_ortho_vert.png"), OB_coef_plot_vert,
          base_width = 5, base_height = 12)
```

```{r fig.height=5, fig.width=14, message=FALSE, warning=FALSE}
OB_coef_plot_hor = ggpubr::ggarrange(BMI_coef_plot, WC_coef_plot, BF_coef_plot,
                  ncol = 3,
                  align = "hv", labels = c("A", "B", "C"))

OB_coef_plot_hor

save_plot(file.path(fig_dir, "OB_coef_plot_hor_ortho.png"), OB_coef_plot_hor,
          base_height = 5, base_width = 12)
```


```{r}
cor.test(Maaslin_OB_BMI_wide$coef_chs_class, Maaslin_OB_BMI_wide$coef_BMI, method = "spearman")
cor.test(Maaslin_OB_WC_wide$coef_chs_class, Maaslin_OB_WC_wide$coef_Waist_Circumference, method = "spearman")
cor.test(Maaslin_OB_BF_wide$coef_chs_class, Maaslin_OB_BF_wide$coef_Fat_percentage, method = "spearman")

```
